requirements.txt

Babel==2.8.0
boto3==1.13.19
botocore==1.16.19
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers
Django
django-extra-views==0.11.0
django-haystack==2.8.1
django-rest-framework==0.1.0
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
docopt==0.6.2
factory-boy==2.12.0
Faker==4.0.3
firebase-admin==4.2.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
django-environ==0.4.5
# Pillow==7.1.2
purl==1.5
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
yarg==0.1.9
